kernel: Move non-kernel db cache size constants
Summary:
These have nothing to do with the txdb, so move them out and into the
node caches.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/31483 | core#31483]]
https://github.com/bitcoin/bitcoin/pull/31483/commits/8826cae285490439dc1f19b25fa70b2b9e62dfe8
Depends on D18581

Note: the renaming of constants was already done in D8318. It's a good idea to keep the `_MB` suffix for now, they will be removed in a later commit when units are converted to bytes.

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18582

+++ src/node/caches.h
@@ -14,8 +14,8 @@ class ArgsManager;
 
 namespace node {
 struct IndexCacheSizes {
-    int64_t tx_index{0};
-    int64_t filter_index{0};
+    size_t tx_index{0};
+    size_t filter_index{0};
 };
 struct CacheSizes {
     IndexCacheSizes index;
